+++ src/components/SelectToken/index.tsx
@@ -89,7 +89,6 @@ const SelectToken = ({
           console.log('searchTokens contract not found addr=', address);
         }
       }
-      console.log('SEEL TTTT=', tokenSearchRes);
       setFoundSearchTokens(tokenSearchRes);
     }
