+++ src/graphql/index.ts
@@ -1,2 +1,3 @@
 export * from './apollo';
 export * from './gql';
+export * from './evmEvents';
